3.0 Update for the Cop Lights
Added more checks for the code to see if the vehicle is alive, if it isn't then it kills the lights. Added a slight Ambient Light Color to the code to give it more realism to real Emergency Lights. If I ever get back on a test server I'll work on the Jeep lighting positions.